quite a lot of updates

+++ app/db/models/summaries.py
@@ -18,4 +18,4 @@ class TextComparison(Base):
     text_id_1 = Column(Integer, nullable=False)
     text_id_2 = Column(Integer, nullable=False)
     user_id = Column(Integer, nullable=False)
-    choice = Column(Integer(2), nullable=False) # Not sure if this 2 makes sense to use here
+    choice = Column(Integer, nullable=False) # Not sure if this 2 makes sense to use here
